Update project, add gateway_event_listeners.{cc,h}
gateway_event_listeners contains event_listener namespace which defines
a base type of objects that gateway push out events to. This is
essentially the same a the Callable function type before, but now has a
defined type (cmd::discord::gateway::event_listener::base). It is kind
of verbose at the moment, maybe change this in the future.

Any classes that wish to receive events must extend this base class and
implement the handle function.