Update gateway,
Changed gateway interface for adding event listeners. Event listener now
subscribe to gateway events and give a name to identify the listener.
This is to allow for more efficient lookups of handlers when getting
specific gateway events. The 'ALL' event can be requested, in which case
the listener is notified of all events.

Start working on support for resuming connections, but this is untested
and the reference is confusing when describing how connections should be
resumed.

The voice gateway interface from the normal gateway is still incomplete.
Probably will add some sort of pipe mechanism to communicate with voice
gateways.