Update voice_state_listener; threaded voice gateway
After successfully being notified by VOICE_SERVER_UPDATE,
voice_state_listener will create a thread to poll events from
the voice_gateway. On destruction of the listener, notifies the thread
to stop and joins it. At the moment there is no forceful method to
make the thread stop if it is blocked on a next_event() call.