Add SSL support for ASIO websocket, fix bug
Fix a bug where the callback of an async_next_message would result in
the WebSocket parser reading the current data being passed to the
completion handler because it was not consumed before the handler was
ran. Now the data is copied into a temporary vector before being handled
to the completion handler.