Move gateway opcodes to discord.h, start music pipeline
audio_decoding is now going to be done through the this program
directly, instead of piping the results of the audio file (from
youtube-dl) into an ffmpeg process.

So far the music streaming works sometimes, but not other times. Also
there is a lot of delay when demuxing, decoding, encoding, encrypting
every time more samples are needed. Probably will end up just decoding
and encoding it all in one go before sending anything.