Move most json parsing to discord.{h,cc}, add audio_source
json usage was scattered around in code. Move a lot of the common
methods to discord.h,cc and added custom structs for easier usage.

audio_source will be the new interface to getting audio data from
multiple sources. For example, current we are using youtube-dl to
retrieve audio from the internet. This might become youtube_source. In
the future there can be many sources that can automatically be used as
long as it implements the next() function to deliver an audio_frame.