refactor code, improve audio source interface
audio_frame is renamed to opus_frame to better indicate that it only is
intended to hold opus data.

audio_resampler::resample changed to return the number of frames, now
takes a void** to point to data

also youtube_dl source is now converting to float instead of s16 because
it seems ffmpeg is using FLTP internally

audio_decoder keeps track of eof and now returns av_frame which contains
the old AVFrame* and also an eof field now to check if decoder has been
flushed.